Change the type of threshold destination distance variable

+++ AD-EYE/ROS_Packages/src/AD-EYE/src/sequence_goal_node.cpp
@@ -60,7 +60,7 @@ private:
     std::string vehicle_state_status_;
 
     // Thresold destination value
-    double threshold_destination_distance_ = 30.00;
+    int threshold_destination_distance_ = 30;
 
     
     /*!
